Add initial project files
Created python scripts for initial setup of SimCompanies simulation project.
Created setup for the logger, decorators, models, and scripts for database handling, logging in, and the main controller. All these were created to facilitate login, data retrieval and handling, and data logging for the simulation project.

Also added a demo setting file to store all configurations in one place. It allows easy access and changes to any configurations.

Scripts have been added to model database objects for item production and requirement details. These scripts help in the abstraction of the database objects making it easier to handle all database operations.

Lastly, a main controller script was added which will serve as the main entry point for running the simulation.